feat: Separate FUTURE flags detection in `check-future-flags` command
The `check-fixmes` command used to report both FIXME and orphan FUTURE
flags. In the future, all commands will have a `--format` argument
that allows text, CSV or JSON output. Reporting FIXME's and FUTURE
tags in a single CSV file seems hard, because they don't have the same
attributes.

As such, `check-fixmes` now only reports FIXME's. A separate command
`check-future-flags` command reports FUTURE flags. One tool, one job.
