+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = check-oldies
-version = 0.8.11.dev0
+version = 1.0.0.dev0
 description = Warns about unattended branches and FIXME or TODO annotations
 long_description = file: README.rst
 long_description_content_type = text/x-rst
@@ -37,8 +37,9 @@ packages=find:
 
 [options.entry_points]
 console-scripts =
-    check-fixmes = check_oldies.check_fixmes:main
     check-branches = check_oldies.check_branches:main
+    check-fixmes = check_oldies.check_fixmes:main
+    check-future-tags = check_oldies.check_future_tags:main
     forget-me-not = check_oldies.forget_me_not:main
 
 [options.extras_require]
